+++ com/system/repositary/UsersRepositary.java
@@ -1,13 +1,10 @@
 package com.system.repositary;
 
-import java.util.List;
-
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.system.model.Billing;
 import com.system.model.Users;
 
-public interface UsersRepositary extends JpaRepository<Users, Integer>{
+public interface UsersRepositary extends JpaRepository<Users, Long>{
+
